app/src/main/java/com/zjl/iwechatmoments/http/handler/HttpUrlConnectHandler.java
package com.zjl.iwechatmoments.http.handler;

import com.zjl.iwechatmoments.http.constant.HttpState;

import java.io.BufferedReader;
import java.io.ByteArrayOutputStream;
import java.io.IOException;
import java.io.InputStream;
import java.io.InputStreamReader;
import java.net.HttpURLConnection;
import java.net.MalformedURLException;
import java.net.URL;
import java.util.Map;

/**
 * Created by zjl on 18-3-15.
 */

public class HttpUrlConnectHandler implements HttpLoadHandler {
    @Override
    public String get(String urlStr) {
        try {
            URL url = new URL(urlStr);
            HttpURLConnection connection = (HttpURLConnection) url.openConnection();
            if (connection.getResponseCode() != HttpState.OK) {
                return null;
            }
            InputStream in = connection.getInputStream();
            BufferedReader reader = new BufferedReader(new InputStreamReader(in));
            StringBuffer response = new StringBuffer();
            String line;
            while ((line = reader.readLine()) != null) {
                response.append(line);
            }
            reader.close();
            in.close();connection.disconnect();
            return response.toString();
        } catch (MalformedURLException e) {
            e.printStackTrace();
        } catch (IOException e) {
            e.printStackTrace();
        }
        return null;
    }

    @Override
    public String post(String url, Map<String, String> params) {
        // TODO: 18-3-15 add post
        return null;
    }
}
